destroy: aws: move ec2 helper functions to their own file
Hopefully it makes aws.go easier to read
